ABC171D

+++ src/abc/171/ABC171C_OneQuadrillionAndOneDalmaions.kt
@@ -1,7 +1,7 @@
 import kotlin.math.pow
 
 fun main() {
-    var n = readLine()!!.toInt()
+    var n = readLine()!!.toLong()
     val result = mutableListOf<Char>();
 
     for (i in (1..99)) {
@@ -9,7 +9,7 @@ fun main() {
         if (n <= pow) {
             n -= 1
             for (j in (1..i)) {
-                result.add('a' + n % 26)
+                result.add('a' + (n % 26).toInt())
                 n = n / 26
             }
             break
